+++ test/expr.test.ts
@@ -121,8 +121,12 @@ describe('Expression', () => {
         });
 
         it('Int8', () => {
+            assert.equal(Expression.Int8.type, 'Int8');
+            assert.equal(Expression.Int8.allocator, Int8Array);
             const idi8 = new Expression.Int8('(a + b) / 2', ['a', 'b']);
             assert.instanceOf(idi8, Expression.Expression);
+            assert.equal(idi8.type, 'Int8');
+            assert.equal(idi8.allocator, Int8Array);
             const ai8 = new Int8Array([1, 2, 3, 4]);
             const ri8 = idi8.map(ai8, 'a', 1);
             assert.instanceOf(ri8, Int8Array);
@@ -131,8 +135,12 @@ describe('Expression', () => {
         });
 
         it('Uint8', () => {
+            assert.equal(Expression.Uint8.type, 'Uint8');
+            assert.equal(Expression.Uint8.allocator, Uint8Array);
             const idui8 = new Expression.Uint8('(a + b) / 2', ['a', 'b']);
             assert.instanceOf(idui8, Expression.Expression);
+            assert.equal(idui8.type, 'Uint8');
+            assert.equal(idui8.allocator, Uint8Array);
             const aui8 = new Uint8Array([1, 2, 3, 4]);
             const rui8 = idui8.map(aui8, 'a', 1);
             assert.instanceOf(rui8, Uint8Array);
@@ -141,8 +149,12 @@ describe('Expression', () => {
         });
 
         it('Int16', () => {
+            assert.equal(Expression.Int16.type, 'Int16');
+            assert.equal(Expression.Int16.allocator, Int16Array);
             const idi16 = new Expression.Int16('(a + b) / 2', ['a', 'b']);
             assert.instanceOf(idi16, Expression.Expression);
+            assert.equal(idi16.type, 'Int16');
+            assert.equal(idi16.allocator, Int16Array);
             const ai16 = new Int16Array([1, 2, 3, 4]);
             const ri16 = idi16.map(ai16, 'a', 1);
             assert.instanceOf(ri16, Int16Array);
@@ -151,8 +163,12 @@ describe('Expression', () => {
         });
 
         it('Uint16', () => {
+            assert.equal(Expression.Uint16.type, 'Uint16');
+            assert.equal(Expression.Uint16.allocator, Uint16Array);
             const idui16 = new Expression.Uint16('(a + b) / 2', ['a', 'b']);
             assert.instanceOf(idui16, Expression.Expression);
+            assert.equal(idui16.type, 'Uint16');
+            assert.equal(idui16.allocator, Uint16Array);
             const aui16 = new Uint16Array([1, 2, 3, 4]);
             const rui16 = idui16.map(aui16, 'a', 1);
             assert.instanceOf(rui16, Uint16Array);
@@ -161,8 +177,12 @@ describe('Expression', () => {
         });
 
         it('Int32', () => {
+            assert.equal(Expression.Int32.type, 'Int32');
+            assert.equal(Expression.Int32.allocator, Int32Array);
             const idi32 = new Expression.Int32('(a + b) / 2', ['a', 'b']);
             assert.instanceOf(idi32, Expression.Expression);
+            assert.equal(idi32.type, 'Int32');
+            assert.equal(idi32.allocator, Int32Array);
             const ai32 = new Int32Array([1, 2, 3, 4]);
             const ri32 = idi32.map(ai32, 'a', 1);
             assert.instanceOf(ri32, Int32Array);
@@ -171,8 +191,12 @@ describe('Expression', () => {
         });
 
         it('Uint32', () => {
+            assert.equal(Expression.Uint32.type, 'Uint32');
+            assert.equal(Expression.Uint32.allocator, Uint32Array);
             const idui32 = new Expression.Uint32('(a + b) / 2', ['a', 'b']);
             assert.instanceOf(idui32, Expression.Expression);
+            assert.equal(idui32.type, 'Uint32');
+            assert.equal(idui32.allocator, Uint32Array);
             const aui32 = new Uint32Array([1, 2, 3, 4]);
             const rui32 = idui32.map(aui32, 'a', 1);
             assert.instanceOf(rui32, Uint32Array);
@@ -181,8 +205,12 @@ describe('Expression', () => {
         });
 
         it('Float32', () => {
+            assert.equal(Expression.Float32.type, 'Float32');
+            assert.equal(Expression.Float32.allocator, Float32Array);
             const id32 = new Expression.Float32('(a + b) / 2', ['a', 'b']);
             assert.instanceOf(id32, Expression.Expression);
+            assert.equal(id32.type, 'Float32');
+            assert.equal(id32.allocator, Float32Array);
             const a32 = new Float32Array([1, 2, 3, 4]);
             const r32 = id32.map(a32, 'a', 1);
             assert.instanceOf(r32, Float32Array);
@@ -191,8 +219,12 @@ describe('Expression', () => {
         });
 
         it('Float64', () => {
+            assert.equal(Expression.Float64.type, 'Float64');
+            assert.equal(Expression.Float64.allocator, Float64Array);
             const id64 = new Expression.Float64('(a + b) / 2', ['a', 'b']);
             assert.instanceOf(id64, Expression.Expression);
+            assert.equal(id64.type, 'Float64');
+            assert.equal(id64.allocator, Float64Array);
             const a64 = new Float64Array([1, 2, 3, 4]);
             const r64 = id64.map(a64, 'a', 1);
             assert.instanceOf(r64, Float64Array);